+++ util/rbac/rbac.go
@@ -11,7 +11,6 @@ import (
 	"time"
 
 	"github.com/argoproj/argo-cd/v3/util/assets"
-	claimsutil "github.com/argoproj/argo-cd/v3/util/claims"
 	"github.com/argoproj/argo-cd/v3/util/glob"
 	jwtutil "github.com/argoproj/argo-cd/v3/util/jwt"
 
@@ -341,7 +340,7 @@ func (e *Enforcer) EnforceErr(rvals ...any) error {
 			if s, ok := rvals[0].(jwt.Claims); ok {
 				claims, err := jwtutil.MapClaims(s)
 				if err == nil {
-					userId := claimsutil.GetUserIdentifier(claims)
+					userId := jwtutil.GetUserIdentifier(claims)
 					if userId != "" {
 						rvalsStrs = append(rvalsStrs, "sub: "+userId)
 					}
